added custom matcher closes #5
PR includes also a little bit of refactoring to the matching logic.
it looks like it adds complexity - but it's actually not.
instead of saving the keys in a dedicated array and looping on it with `index of` - I loop over the original array using the custom comparer function.
also - using a simple string creates a custom comparer that use that key.